Decrease the size of the semantic class voxel data
Most detectors and datasets so far use fewer than 100 classes so 255
should be enough for now.

+++ include/se/map/utils/type_util.hpp
@@ -57,7 +57,7 @@ typedef float time_stamp_t; ///< The type of the time stamp
 
 typedef uint32_t rgba_t; ///< The type of the colour
 
-typedef short semantics_t; ///< The type of the semantic class
+typedef uint8_t semantics_t; ///< The type of the semantic class
 
 } // namespace se
